+++ CHANGELOG.md
@@ -17,7 +17,8 @@
 -----------
 
 ### Internals
-* None.
+* Add admin api and test for performing the PBS->FLX migration and roll back on the server. (PR [#6366](https://github.com/realm/realm-core/pull/6366))
+* Integrate protocol support for PBS->FLX client migration ([PR #6355](https://github.com/realm/realm-core/pull/6355))
 
 ----------------------------------------------
 
@@ -38,8 +39,7 @@
 -----------
 
 ### Internals
-* Add admin api and test for performing the PBS->FLX migration and roll back on the server. (PR [#6366](https://github.com/realm/realm-core/pull/6366))
-
+* None
 ----------------------------------------------
 
 # 13.7.0 Release notes